refactor: move calc functions into picked shapes file
